requirements.txt
fastapi==0.115.12
uvicorn==0.23.2
pydantic==2.11.3
httpx==0.25.0
beautifulsoup4==4.12.2
lxml==5.3.2
python-dotenv==1.0.0
tenacity==8.2.3
pytest==7.4.3
requests==2.32.0
aiohttp==3.10.11
selenium==4.15.2
webdriver-manager==4.0.1
playwright==1.51.0
fake-useragent==1.3.0
greenlet==3.2.0